Add base for skinny client with tests
