+++ tests/conftest.py
@@ -3,7 +3,6 @@ import os
 import pytest
 
 import mlflow
-from mlflow.utils.file_utils import path_to_local_sqlite_uri
 
 
 @pytest.fixture
@@ -23,6 +22,8 @@ def reset_mock():
 
 @pytest.fixture(autouse=True)
 def tracking_uri_mock(tmpdir, request):
+    from mlflow.utils.file_utils import path_to_local_sqlite_uri
+
     try:
         if "notrackingurimock" not in request.keywords:
             tracking_uri = path_to_local_sqlite_uri(os.path.join(tmpdir.strpath, "mlruns"))
@@ -33,3 +34,9 @@ def tracking_uri_mock(tmpdir, request):
         mlflow.set_tracking_uri(None)
         if "notrackingurimock" not in request.keywords:
             del os.environ["MLFLOW_TRACKING_URI"]
+
+
+@pytest.fixture
+def reset_active_experiment():
+    yield
+    mlflow.tracking.fluent._active_experiment_id = None
