+++ tests/tracking/conftest.py
@@ -2,8 +2,6 @@ import os
 
 import pytest
 
-import mlflow
-
 
 @pytest.fixture
 def tmp_wkdir(tmpdir):
@@ -11,9 +9,3 @@ def tmp_wkdir(tmpdir):
     os.chdir(str(tmpdir))
     yield
     os.chdir(initial_wkdir)
-
-
-@pytest.fixture
-def reset_active_experiment():
-    yield
-    mlflow.tracking.fluent._active_experiment_id = None
