+++ tests/tracking/registry/test_unit_model_registry_client.py
@@ -1,28 +1,14 @@
 import pytest
 from unittest import mock
 
-from mlflow.entities import SourceType, ViewType, RunTag, Run, RunInfo
+from mlflow.entities import Run, RunInfo
 from mlflow.entities.model_registry import ModelVersion, ModelVersionTag
 from mlflow.exceptions import MlflowException
 from mlflow.protos.databricks_pb2 import ErrorCode, FEATURE_DISABLED
-from mlflow.store.tracking import SEARCH_MAX_RESULTS_DEFAULT
-from mlflow.tracking import set_registry_uri, MlflowClient
+from mlflow.tracking import MlflowClient
 from mlflow.utils.file_utils import TempDir
-from mlflow.utils.mlflow_tags import (
-    MLFLOW_USER,
-    MLFLOW_SOURCE_NAME,
-    MLFLOW_SOURCE_TYPE,
-    MLFLOW_PARENT_RUN_ID,
-    MLFLOW_GIT_COMMIT,
-    MLFLOW_PROJECT_ENTRY_POINT,
-)
+from mlflow.tracking import set_registry_uri
 from mlflow.utils.uri import construct_run_url
-
-
-@pytest.fixture
-def mock_store():
-    with mock.patch("mlflow.tracking._tracking_service.utils._get_store") as mock_get_store:
-        yield mock_get_store.return_value
 
 
 @pytest.fixture
@@ -31,149 +17,8 @@ def mock_registry_store():
         yield mock_get_store.return_value
 
 
-@pytest.fixture
-def mock_time():
-    time = 1552319350.244724
-    with mock.patch("time.time", return_value=time):
-        yield time
-
-
-def test_client_create_run(mock_store, mock_time):
-    experiment_id = mock.Mock()
-
-    MlflowClient().create_run(experiment_id)
-
-    mock_store.create_run.assert_called_once_with(
-        experiment_id=experiment_id, user_id="unknown", start_time=int(mock_time * 1000), tags=[],
-    )
-
-
-def test_client_create_run_overrides(mock_store):
-    experiment_id = mock.Mock()
-    user = mock.Mock()
-    start_time = mock.Mock()
-    tags = {
-        MLFLOW_USER: user,
-        MLFLOW_PARENT_RUN_ID: mock.Mock(),
-        MLFLOW_SOURCE_TYPE: SourceType.to_string(SourceType.JOB),
-        MLFLOW_SOURCE_NAME: mock.Mock(),
-        MLFLOW_PROJECT_ENTRY_POINT: mock.Mock(),
-        MLFLOW_GIT_COMMIT: mock.Mock(),
-        "other-key": "other-value",
-    }
-
-    MlflowClient().create_run(experiment_id, start_time, tags)
-
-    mock_store.create_run.assert_called_once_with(
-        experiment_id=experiment_id,
-        user_id=user,
-        start_time=start_time,
-        tags=[RunTag(key, value) for key, value in tags.items()],
-    )
-    mock_store.reset_mock()
-    MlflowClient().create_run(experiment_id, start_time, tags)
-    mock_store.create_run.assert_called_once_with(
-        experiment_id=experiment_id,
-        user_id=user,
-        start_time=start_time,
-        tags=[RunTag(key, value) for key, value in tags.items()],
-    )
-
-
-def test_client_search_runs_defaults(mock_store):
-    MlflowClient().search_runs([1, 2, 3])
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=[1, 2, 3],
-        filter_string="",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_filter(mock_store):
-    MlflowClient().search_runs(["a", "b", "c"], "my filter")
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=["a", "b", "c"],
-        filter_string="my filter",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_view_type(mock_store):
-    MlflowClient().search_runs(["a", "b", "c"], "my filter", ViewType.DELETED_ONLY)
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=["a", "b", "c"],
-        filter_string="my filter",
-        run_view_type=ViewType.DELETED_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_max_results(mock_store):
-    MlflowClient().search_runs([5], "my filter", ViewType.ALL, 2876)
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=[5],
-        filter_string="my filter",
-        run_view_type=ViewType.ALL,
-        max_results=2876,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_int_experiment_id(mock_store):
-    MlflowClient().search_runs(123)
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=[123],
-        filter_string="",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_string_experiment_id(mock_store):
-    MlflowClient().search_runs("abc")
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=["abc"],
-        filter_string="",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token=None,
-    )
-
-
-def test_client_search_runs_order_by(mock_store):
-    MlflowClient().search_runs([5], order_by=["a", "b"])
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=[5],
-        filter_string="",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=["a", "b"],
-        page_token=None,
-    )
-
-
-def test_client_search_runs_page_token(mock_store):
-    MlflowClient().search_runs([5], page_token="blah")
-    mock_store.search_runs.assert_called_once_with(
-        experiment_ids=[5],
-        filter_string="",
-        run_view_type=ViewType.ACTIVE_ONLY,
-        max_results=SEARCH_MAX_RESULTS_DEFAULT,
-        order_by=None,
-        page_token="blah",
-    )
+def _default_model_version():
+    return ModelVersion("model name", 1, creation_timestamp=123, status="READY")
 
 
 def test_client_registry_operations_raise_exception_with_unsupported_registry_store():
@@ -271,26 +116,6 @@ def test_registry_uri_from_set_registry_uri():
     client = MlflowClient(tracking_uri="databricks://tracking")
     assert client._registry_uri == uri
     set_registry_uri(None)
-
-
-def test_registry_uri_from_tracking_uri_param():
-    with mock.patch(
-        "mlflow.tracking._tracking_service.utils.get_tracking_uri"
-    ) as get_tracking_uri_mock:
-        get_tracking_uri_mock.return_value = "databricks://default_tracking"
-        tracking_uri = "databricks://tracking_vhawoierj"
-        client = MlflowClient(tracking_uri=tracking_uri)
-        assert client._registry_uri == tracking_uri
-
-
-def test_registry_uri_from_implicit_tracking_uri():
-    with mock.patch(
-        "mlflow.tracking._tracking_service.utils.get_tracking_uri"
-    ) as get_tracking_uri_mock:
-        tracking_uri = "databricks://tracking_wierojasdf"
-        get_tracking_uri_mock.return_value = tracking_uri
-        client = MlflowClient()
-        assert client._registry_uri == tracking_uri
 
 
 def test_create_model_version_nondatabricks_source_no_runlink(mock_registry_store):
@@ -442,5 +267,21 @@ def test_create_model_version_copy_not_called_to_nondb(mock_registry_store):
         upload_mock.assert_not_called()
 
 
-def _default_model_version():
-    return ModelVersion("model name", 1, creation_timestamp=123, status="READY")
+def test_registry_uri_from_tracking_uri_param():
+    with mock.patch(
+        "mlflow.tracking._tracking_service.utils.get_tracking_uri"
+    ) as get_tracking_uri_mock:
+        get_tracking_uri_mock.return_value = "databricks://default_tracking"
+        tracking_uri = "databricks://tracking_vhawoierj"
+        client = MlflowClient(tracking_uri=tracking_uri)
+        assert client._registry_uri == tracking_uri
+
+
+def test_registry_uri_from_implicit_tracking_uri():
+    with mock.patch(
+        "mlflow.tracking._tracking_service.utils.get_tracking_uri"
+    ) as get_tracking_uri_mock:
+        tracking_uri = "databricks://tracking_wierojasdf"
+        get_tracking_uri_mock.return_value = tracking_uri
+        client = MlflowClient()
+        assert client._registry_uri == tracking_uri
